+++ src/core/constraint_template.jl
@@ -350,3 +350,25 @@ function constraint_mc_transformer_power_traditional_on_off(pm::AbstractUnbalanc
         error("Zig-zag not yet supported.")
     end
 end
+
+
+"""
+    constraint_grid_forming_inverter_per_cc_block(pm::AbstractUnbalancedPowerModel; nw::Int=nw_id_default, relax::Bool=false)
+
+Template function for constraining the number of grid-forming inverters per connected component in the block mld problem
+"""
+function constraint_grid_forming_inverter_per_cc_block(pm::AbstractUnbalancedPowerModel; nw::Int=nw_id_default, relax::Bool=false)
+    constraint_grid_forming_inverter_per_cc_block(pm, nw; relax=relax)
+end
+
+
+"""
+    constraint_grid_forming_inverter_per_cc_traditional(pm::AbstractUnbalancedPowerModel; nw::Int=nw_id_default, relax::Bool=false)
+
+Template function for constraining the number of grid-forming inverters per connected component in the traditional mld formulation
+"""
+function constraint_grid_forming_inverter_per_cc_traditional(pm::AbstractUnbalancedPowerModel; nw::Int=nw_id_default, relax::Bool=false)
+    constraint_grid_forming_inverter_per_cc_traditional(pm, nw; relax=relax)
+end
+
+